Move policy helper functions into a separate file
This will help keep things organized.

(cherry picked from commit 6a68d62ce7157b0d602f1bc24acd8a3ca2b4fc43)
